requirements.txt
# LangChain 0.3.x — supports Python 3.13 + numpy 2.x
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-groq==0.3.2
langchain-huggingface==0.1.2

# Vector store
faiss-cpu==1.9.0.post1

# Embeddings (local, free)
sentence-transformers==3.4.1

# Document loading
beautifulsoup4==4.12.3
requests==2.32.3
pypdf==4.3.1
lxml==5.3.0

# UI
streamlit==1.43.2

# Utils
numpy==2.1.3
pandas==2.2.3
tqdm==4.66.5
python-dotenv==1.0.1
tiktoken==0.8.0